readarpa
Does not yet work. The idea is to mmap the whole file and build vectors of doubles and strings to pass to the hash constructor. The process is simple in C, but I need to learn the .Call interface to R. This will not be portable, as Windows does not support mmap, so hopefully someone that knows Windows can eventually help port it if this package proves useful.